restaurant/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-debug-toolbar==4.1.0
django-model-utils==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
